RAG/CodeBot/Frontend/requirements.txt
streamlit-code-editor
streamlit>=1.32
requests
pandas
